Implemented 5 min cooling period usecase

+++ Models/OrderModel.go
@@ -1,10 +1,10 @@
 package Models
 
 type Order struct {
-	OrderID        int16 `json:"order_id" gorm:"primary_key"`
-	ProductProductID int16 `json:"product_id"`
+	OrderID        uint `json:"order_id" gorm:"primary_key"`
+	ProductProductID uint `json:"product_id"`
 	Quantity  int16  `json:"qty"`
-	CustomerCustomerID int16 `json:"customer_id"`
+	CustomerCustomerID uint `json:"customer_id"`
 	Status     string `json:"status"`
 	Product Product
 	Customer Customer
