+++ Models/ProductModel.go
@@ -1,7 +1,7 @@
 package Models
 
 type Product struct {
-	ProductID int16 `json:"product_id" gorm:"primary_key"`
+	ProductID uint `json:"product_id" gorm:"primary_key"`
 	Name      string `json:"name"`
 	Price     int16  `json:"price"`
 	Quantity  int16  `json:"qty"`
